+++ cpp/src/arrow/engine/CMakeLists.txt
@@ -28,7 +28,8 @@ set(ARROW_SUBSTRAIT_SRCS
     substrait/serde.cc
     substrait/test_plan_builder.cc
     substrait/type_internal.cc
-    substrait/util.cc)
+    substrait/util.cc
+    substrait/util_internal.cc)
 
 add_arrow_lib(arrow_substrait
               CMAKE_PACKAGE_NAME
