+++ cpp/src/arrow/engine/substrait/util.cc
@@ -162,15 +162,6 @@ const std::string& default_extension_types_uri() {
   return uri;
 }
 
-std::string EnumToString(int value, const google::protobuf::EnumDescriptor* descriptor) {
-  const google::protobuf::EnumValueDescriptor* value_desc =
-      descriptor->FindValueByNumber(value);
-  if (value_desc == nullptr) {
-    return "unknown";
-  }
-  return value_desc->name();
-}
-
 }  // namespace engine
 
 }  // namespace arrow
